third_party/blink/web_tests/external/wpt/html/rendering/widgets/button-layout/computed-style-expected.txt

This is a testharness.js-based test.
Found 78 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] computed display of <input type=reset> with display: inline
  assert_equals: expected "inline" but got "inline-block"
[FAIL] computed display of <input type=button> with display: inline
  assert_equals: expected "inline" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: inline
  assert_equals: expected "inline" but got "inline-block"
[FAIL] computed display of <input type=color> with display: inline
  assert_equals: expected "inline" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: inline
  assert_equals: expected "inline" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: list-item
  assert_equals: expected "list-item" but got "block"
[FAIL] computed display of <input type=button> with display: list-item
  assert_equals: expected "list-item" but got "block"
[FAIL] computed display of <input type=submit> with display: list-item
  assert_equals: expected "list-item" but got "block"
[FAIL] computed display of <input type=color> with display: list-item
  assert_equals: expected "list-item" but got "block"
[FAIL] computed display of <button type=submit> with display: list-item
  assert_equals: expected "list-item" but got "block"
[FAIL] computed display of <input type=reset> with display: table
  assert_equals: expected "table" but got "block"
[FAIL] computed display of <input type=button> with display: table
  assert_equals: expected "table" but got "block"
[FAIL] computed display of <input type=submit> with display: table
  assert_equals: expected "table" but got "block"
[FAIL] computed display of <input type=color> with display: table
  assert_equals: expected "table" but got "block"
[FAIL] computed display of <button type=submit> with display: table
  assert_equals: expected "table" but got "block"
[FAIL] computed display of <input type=reset> with display: inline-table
  assert_equals: expected "inline-table" but got "inline-block"
[FAIL] computed display of <input type=button> with display: inline-table
  assert_equals: expected "inline-table" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: inline-table
  assert_equals: expected "inline-table" but got "inline-block"
[FAIL] computed display of <input type=color> with display: inline-table
  assert_equals: expected "inline-table" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: inline-table
  assert_equals: expected "inline-table" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: table-row-group
  assert_equals: expected "table-row-group" but got "inline-block"
[FAIL] computed display of <input type=button> with display: table-row-group
  assert_equals: expected "table-row-group" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: table-row-group
  assert_equals: expected "table-row-group" but got "inline-block"
[FAIL] computed display of <input type=color> with display: table-row-group
  assert_equals: expected "table-row-group" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: table-row-group
  assert_equals: expected "table-row-group" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: table-header-group
  assert_equals: expected "table-header-group" but got "inline-block"
[FAIL] computed display of <input type=button> with display: table-header-group
  assert_equals: expected "table-header-group" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: table-header-group
  assert_equals: expected "table-header-group" but got "inline-block"
[FAIL] computed display of <input type=color> with display: table-header-group
  assert_equals: expected "table-header-group" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: table-header-group
  assert_equals: expected "table-header-group" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: table-footer-group
  assert_equals: expected "table-footer-group" but got "inline-block"
[FAIL] computed display of <input type=button> with display: table-footer-group
  assert_equals: expected "table-footer-group" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: table-footer-group
  assert_equals: expected "table-footer-group" but got "inline-block"
[FAIL] computed display of <input type=color> with display: table-footer-group
  assert_equals: expected "table-footer-group" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: table-footer-group
  assert_equals: expected "table-footer-group" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: table-row
  assert_equals: expected "table-row" but got "inline-block"
[FAIL] computed display of <input type=button> with display: table-row
  assert_equals: expected "table-row" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: table-row
  assert_equals: expected "table-row" but got "inline-block"
[FAIL] computed display of <input type=color> with display: table-row
  assert_equals: expected "table-row" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: table-row
  assert_equals: expected "table-row" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: table-column-group
  assert_equals: expected "table-column-group" but got "inline-block"
[FAIL] computed display of <input type=button> with display: table-column-group
  assert_equals: expected "table-column-group" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: table-column-group
  assert_equals: expected "table-column-group" but got "inline-block"
[FAIL] computed display of <input type=color> with display: table-column-group
  assert_equals: expected "table-column-group" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: table-column-group
  assert_equals: expected "table-column-group" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: table-column
  assert_equals: expected "table-column" but got "inline-block"
[FAIL] computed display of <input type=button> with display: table-column
  assert_equals: expected "table-column" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: table-column
  assert_equals: expected "table-column" but got "inline-block"
[FAIL] computed display of <input type=color> with display: table-column
  assert_equals: expected "table-column" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: table-column
  assert_equals: expected "table-column" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: table-cell
  assert_equals: expected "table-cell" but got "inline-block"
[FAIL] computed display of <input type=button> with display: table-cell
  assert_equals: expected "table-cell" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: table-cell
  assert_equals: expected "table-cell" but got "inline-block"
[FAIL] computed display of <input type=color> with display: table-cell
  assert_equals: expected "table-cell" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: table-cell
  assert_equals: expected "table-cell" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: table-caption
  assert_equals: expected "table-caption" but got "inline-block"
[FAIL] computed display of <input type=button> with display: table-caption
  assert_equals: expected "table-caption" but got "inline-block"
[FAIL] computed display of <input type=submit> with display: table-caption
  assert_equals: expected "table-caption" but got "inline-block"
[FAIL] computed display of <input type=color> with display: table-caption
  assert_equals: expected "table-caption" but got "inline-block"
[FAIL] computed display of <button type=submit> with display: table-caption
  assert_equals: expected "table-caption" but got "inline-block"
[FAIL] computed display of <input type=reset> with display: ruby-base
  assert_not_equals: display: ruby-base is not supported got disallowed value ""
[FAIL] computed display of <input type=button> with display: ruby-base
  assert_not_equals: display: ruby-base is not supported got disallowed value ""
[FAIL] computed display of <input type=submit> with display: ruby-base
  assert_not_equals: display: ruby-base is not supported got disallowed value ""
[FAIL] computed display of <input type=color> with display: ruby-base
  assert_not_equals: display: ruby-base is not supported got disallowed value ""
[FAIL] computed display of <input type=image> with display: ruby-base
  assert_not_equals: display: ruby-base is not supported got disallowed value ""
[FAIL] computed display of <button type=submit> with display: ruby-base
  assert_not_equals: display: ruby-base is not supported got disallowed value ""
[FAIL] computed display of <input type=reset> with display: ruby-base-container
  assert_not_equals: display: ruby-base-container is not supported got disallowed value ""
[FAIL] computed display of <input type=button> with display: ruby-base-container
  assert_not_equals: display: ruby-base-container is not supported got disallowed value ""
[FAIL] computed display of <input type=submit> with display: ruby-base-container
  assert_not_equals: display: ruby-base-container is not supported got disallowed value ""
[FAIL] computed display of <input type=color> with display: ruby-base-container
  assert_not_equals: display: ruby-base-container is not supported got disallowed value ""
[FAIL] computed display of <input type=image> with display: ruby-base-container
  assert_not_equals: display: ruby-base-container is not supported got disallowed value ""
[FAIL] computed display of <button type=submit> with display: ruby-base-container
  assert_not_equals: display: ruby-base-container is not supported got disallowed value ""
[FAIL] computed display of <input type=reset> with display: ruby-text-container
  assert_not_equals: display: ruby-text-container is not supported got disallowed value ""
[FAIL] computed display of <input type=button> with display: ruby-text-container
  assert_not_equals: display: ruby-text-container is not supported got disallowed value ""
[FAIL] computed display of <input type=submit> with display: ruby-text-container
  assert_not_equals: display: ruby-text-container is not supported got disallowed value ""
[FAIL] computed display of <input type=color> with display: ruby-text-container
  assert_not_equals: display: ruby-text-container is not supported got disallowed value ""
[FAIL] computed display of <input type=image> with display: ruby-text-container
  assert_not_equals: display: ruby-text-container is not supported got disallowed value ""
[FAIL] computed display of <button type=submit> with display: ruby-text-container
  assert_not_equals: display: ruby-text-container is not supported got disallowed value ""
Harness: the test ran to completion.
